+++ apps/pushnotif-func/src/domain/push-service.ts
@@ -1,5 +1,13 @@
+import { ErrorInternal, ErrorNotFound, ErrorTooManyRequests } from "./error";
 import { JsonPatch } from "./json-patch";
 
 export interface InstallationRepository {
-  updateInstallation(id: string, patches: JsonPatch[]): Promise<Error | string>;
+  /**
+   * updateInstallation perform an update operation on the installation
+   * identified by the `id` using the provided patches.
+   * */
+  updateInstallation(
+    id: string,
+    patches: JsonPatch[],
+  ): Promise<ErrorInternal | ErrorNotFound | ErrorTooManyRequests | string>;
 }