+++ apps/cqrs-func/CHANGELOG.md
@@ -1,5 +1,11 @@
 # cqrs-func
 
+## 2.0.0
+
+### Major Changes
+
+- 9479207: Migrate cqrs-func to Programming Model v4
+
 ## 1.0.9
 
 ### Patch Changes