+++ apps/services-func/CHANGELOG.md
@@ -1,5 +1,17 @@
 # services-func
 
+## 2.0.2
+
+### Patch Changes
+
+- ea8c8f8: remove azure-storage-legacy-migration-kit
+
+## 2.0.1
+
+### Patch Changes
+
+- c4d2e27: Fix applicationInsight Initialization
+
 ## 2.0.0
 
 ### Major Changes